Add vertex normals alongside face normals

+++ model/off/off_document_4d.h
@@ -31,8 +31,8 @@ class OFFDocument4D : public Resource {
 	static String _cell_or_face_to_off_string(const PackedInt32Array &p_face);
 
 	void _count_unique_edges_from_faces();
-	int _find_or_insert_face(const int p_a, const int p_b, const int p_c, const bool p_deduplicate_faces = true);
-	int _find_or_insert_vertex(const Vector4 &p_vertex, const bool p_deduplicate_vertices = true);
+	int64_t _find_or_insert_face(const int32_t p_a, const int32_t p_b, const int32_t p_c, const bool p_deduplicate_faces = true);
+	int64_t _find_or_insert_vertex(const Vector4 &p_vertex, const bool p_deduplicate_vertices = true);
 	TypedArray<PackedInt32Array> _calculate_cell_vertex_indices();
 	TypedArray<PackedInt32Array> _calculate_simplex_vertex_indices(const TypedArray<PackedInt32Array> &p_cell_vertex_indices);
